+++ src/RuleSet/ExplicitRuleSet.php
@@ -20,6 +20,7 @@ namespace Ergebnis\PhpCsFixer\Config\RuleSet;
  *
  * - does not configure any rules for rule sets (@see https://github.com/FriendsOfPHP/PHP-CS-Fixer/blob/v2.17.3/doc/ruleSets/index.rst)
  * - configures a rule that can be configured with an explicit configuration when the rule is enabled
+ * - configures every rule that is not deprecated
  */
 interface ExplicitRuleSet
 {